[refinery] Update to set redis host based on release name (#194)
## Which problem is this PR solving?

Currently the refinery chart must be installed with a release name of
`refinery` or the `.Values.config.PeerManagement.RedisHost` must be
manually updated to match the release name. This works if you know what
the release name will be, but for situations like chart-testing, the
release name is dynamically generated. It is also problematic when using
the chart for the first time; if you want to use the default values.yaml
you must use `refinery` as the release name.

## Short description of the changes

Updates `.Values.config` to be run through the `tpl` function. This
allows us to set `RedisHost: '{{include "refinery.redis.fullname"
.}}:6379'` in the values.yaml and us the chart's fullname helpers to
determine the default RedisHost.

## How to verify that this has the expected result

Deployed refinery local and ran it through chart-testing.

Co-authored-by: Jamie Danielson <jamieedanielson@gmail.com>