[chore] add chart testing (#193)
## Which problem is this PR solving?

Currently all tests of the helm charts have to be done manually. This
can be time consuming and it can be difficult to cover all useful test
scenarios

## Short description of the changes

Adds a new circleci workflow to use
[chart-testing](https://github.com/helm/chart-testing) to test the helm
charts. This allows automatic detection of changed charts and then
installs all the values.yamls in the `ci` folder, one at a time, to
ensure the chart is installable.

## How to verify that this has the expected result

copious amounts of trial and error (118 to be exact) lol