maint: cleanup release (#250)
## Which problem is this PR solving?

With the bump to chart releaser 1.5.0 we lost the ability to do `cr
index` with an `ssh` git remote; the tool now requires us to use an
`https` remote. For the `v1.2.3-secure-tenancy` release I had to do the
`cr index` step manually from my laptop. See and
https://github.com/helm/chart-releaser/issues/124.

The PR solves this problem and also removes some deprecated flags and
unused env vars.

## Short description of the changes

- set git remote to https url for `cr index` step.
- remove unused `--charts-repo` flag
- remove unused variables.

## How to verify that this has the expected result

I couldn't :/ but I'll try to do a honeycomb chart release once this is
merged to test it.