docs: [refinery] Add MaxConnectionAge default to values.yaml (#254)
<!--
Thank you for contributing to the project! 💜
Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Which problem is this PR solving?

The MaxConnectionAge default is 0 which leads to load balancers failing
to re-balance GRPC connections to Refinery. This means that one node in
a multi-node deployment receives all of the traffic and expends more
resources than it should redirecting all the traffic while the other
nodes sit idle.

This PR doesn't change the default but exposes it as a configuration
that can be set so operators can be in a better position to fix the
problem if they run into it.

## Short description of the changes

Adds some new  

## How to verify that this has the expected result

You can use the `helm template` command to see that the config is
present. When it comes to actually verifying the connections remain
open, I haven't cracked that nut yet.