maint: Release on every commit (#309)
## Which problem is this PR solving?

Currently to release a chart we must create a prep PR and then manually
tag. Based on the infrequency of releases and the scope of what is being
released, we now consider this toil.

This PR updates the repo's release process to be in line with other
popular charts like opentelemetry, jaeger, and prometheus. With these
changes a chart will be released on every commit to main.

## Short description of the changes

- Updates the release to identify and release charts that have changed
in the latest commit
- Since we use circleci we coudn't use the chart-releaser-action
directly, so instead we use their bash script.
- Updated docs

## How to verify that this has the expected result