feat(refinery): add deployment-marker support (#400)
## Which problem is this PR solving?

It's helpful to know when a new version of Refinery has been deployed
when viewing Refinery's telemetry data. This PR adds a deployment marker
job in Refinery's chart. The job will create a honeycomb marker when a
Refinery deployment happenes.

NOTE: This is a MVP version of this feature. There could be other
improvements made down the line

## Short description of the changes

- add a deployment-marker job. The job utilize helm hooks to only run
after a `helm install` or `helm upgrade`
- introduce a new `REFINERY_HONEYCOMB_DEPLOY_MARKER_API_KEY` environment
variable for sending deploy markers to honeycomb

## How to verify that this has the expected result
Tested it locally

---------

Co-authored-by: Tyler Helmuth <12352919+TylerHelmuth@users.noreply.github.com>