feat(refinery): add template (tpl) support to extraCommandArgs (#401)
## Which problem is this PR solving?

Refinery supports layering of config/rules and has the capability to get
both from a URL. When using this chart as a subchart it is not possible
to pass in additional config/rules using a url that is based on the
parent chart unless this chart supports templating the input.

## Short description of the changes

Add `tpl` support to `extraCommandArgs`

## How to verify that this has the expected result

Tested locally.  Adding

```yaml
extraCommandArgs:
  - "{{ .Release.Name }}"
```

results in 

```
command:
  - refinery
  - -c
  - /etc/refinery/config.yaml
  - -r
  - /etc/refinery/rules.yaml
  - testing
```

Non-templated strings are unaffected.