feat(observability-pipeline): add control plane telemetry support (#407)
## Which problem is this PR solving?

Adds ability to enable the control plane's instrumentation using otel's
sdk configuration. When enabled, the chart will include and mount a
configmap for configuring the exporting of the control plane's
telemetry.

## Short description of the changes

- add new configmap template
- add conditions to the control plane deployment to hook up config
- add default otel sdk config to push to the gateway collector that can
be installed by this chart. This is the default instead of pushing to
honeycomb because the otel configuraiton doesn't support env vars yet,
so there isn't a way to pass in our honeycomb ingest key as a secret.
Instead the default is to export the data to a collector which can
properly manage the secret (and enhance with k8s attributes).

## How to verify that this has the expected result

Tested locally.