Update gh-pages README to match main (#444)
<!--
Thank you for contributing to the project! 💜
Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Which problem is this PR solving?

The README in the `gh-pages` branch (which displays at
https://honeycombio.github.io/helm-charts/) is out of date.

## Short description of the changes

This makes it current with `main`.

## How to verify that this has the expected result
No PR previews for GitHub Pages! So just have to compare READMEs.

https://github.com/honeycombio/helm-charts/blob/jessica.update-gh-pages/README.md
should look the same as
https://github.com/honeycombio/helm-charts/blob/main/README.md
And the links on both should go to the same places.