feat(refinery): Make it easy to set eu1 urls (#460)
## Which problem is this PR solving?

Right now eu customer need to manually set a couple refinery
configuration endpoints to point to eu honeycomb instead of prod. That
is kind of annoying. This PR allows the chart to automatically set those
fields based on a flag.

## Short description of the changes
- add new `region` setting that will automatically configure refinery's
endpoints based on the value
- Add comments

## How to verify that this has the expected result
tested locally using `helm template`