feat(observability-pipeline): switch collector telemetry config to file (#469)
## Which problem is this PR solving?

closes https://github.com/honeycombio/pipeline-team/issues/255

## Short description of the changes

Switch the primary collector's telemetry settings to a local file from
helm instead of from Beekeeper. This brings it in line with beekeeper,
refinery and the opampsupervisor. It also ensures the collector can emit
telemetry even if it doesn't get the chance to talk to beekeeper

## How to verify that this has the expected result

local testing and templating