feat: Added template rendering tests (#490)
<!--
Thank you for contributing to the project! 💜
Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Which problem is this PR solving?

When committing changes that will change the rendered output, it's hard
to know if there are adverse effects to that.

## Short description of the changes

This PR adds 2 testing scripts.

`render-all-test-templates.sh` which will render all the charts based on
values files located in the tests folder. This also strips the details
that always change like the chart version to make it clearer.
`compare-test-rendered.sh` which will compare 2 sets of rendered files.

I've also added this to CI, so that if someone has not committed the
updated rendered files CI will fail. This will ensure that PR reviews
show all the expected changes a customer will see.

## How to verify that this has the expected result

The build should pass, this PR makes no changes to the charts.