fix(tests): Fixed the test script as it wasn't re-rendering and testing all chart templates (#496)
## Which problem is this PR solving?

Rendering tests had a bug, and were therefore only testing refinery.
This meant that the test-render hadn't picked up a change in the
`securityContext` for redis which is now included in this PR.

## Short description of the changes

fixed the `test-render` action, and added an easy make command for
re-rendering the files.

## How to verify that this has the expected result

Build pipeline doesn't fail