maint(htp-builder): dont subchart refinery (#505)
## Which problem is this PR solving?

Switch htp-builder to install refinery itself instead of relying on the
refinery subchart. This will give us greater control over the refinery
install, especially when trying to configure refinery env vars and other
properties based on root values from the chart.

This refinery install has significantly less features that the refinery
chart - this is intentional. Things like tolerations and annotations are
missing. We know users will want these features - we'll add back
features as `htp-builder` matures, and likely add them to other
artifacts at the same time.

Removing the subchart means we don't need anything in `global` anymore,
and we'll be able to improve refinery's env var support so that we dont
need to put things in `refinery.extraEnvs` that we know we need to add.
I'll do these in followup PRs.

## How to verify that this has the expected result

installed via sippycup