feat: Remove requirement for setting ingest key (#564)
## Which problem is this PR solving?

Updates the HTP chart to not require ingestKey be provided. When not
provided, configures beekeeper to generate ingest keys automatically.

NOTE: This is in draft until the new beekeeper release is available.

## Short description of the changes
- Update htp-builder chart version to 0.2.0
- Update notes.txt to remove ingestkey value or secret requirement
- Set new GENERATE_INGEST_KEYS env var in beekeeper deployment when
neither ingestKey.value or ingestKey.secret is set
- Set HTP_EXPORTER_APIKEY env var in primary-collector and refinery
deployment configs only when ingestKey.value or ingestKey.secret is set

---------

Co-authored-by: Tyler Helmuth <12352919+TylerHelmuth@users.noreply.github.com>